requirements.txt
paho-mqtt==1.5.0
matplotlib==3.2.1